+++ agent/src/main/java/com/teamscale/jacoco/agent/options/ETestwiseCoverageMode.java
@@ -6,6 +6,8 @@ import com.teamscale.jacoco.agent.testimpact.TestEventHandlerStrategyBase;
 public enum ETestwiseCoverageMode {
 	/** Caches testwise coverage in-memory and uploads a report to Teamscale. */
 	TEAMSCALE_UPLOAD,
+	/** Writes testwise coverage to disk as .json files. */
+	DISK,
 	/** Writes testwise coverage to disk as .exec files. */
 	EXEC_FILE,
 	/** Returns testwise coverage to the caller via HTTP. */